Renamed Canvas as SvgDocument, XYChart as Chart, and added Layer
